requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
attrs==20.2.0             # via pytest
backcall==0.2.0           # via ipython
certifi==2020.6.20        # via requests, skyfield
chardet==3.0.4            # via requests
click==7.1.2              # via pip-tools
decorator==4.4.2          # via ipython
freezegun==1.0.0          # via -r requirements.in
idna==2.10                # via requests
importlib-metadata==2.0.0  # via pluggy, pytest
iniconfig==1.0.1          # via pytest
ipdb==0.13.3              # via -r requirements.in
ipython-genutils==0.2.0   # via traitlets
ipython==7.18.1           # via ipdb
jedi==0.17.2              # via ipython
jplephem==2.15            # via skyfield
lunardate==0.2.0          # via workalendar
more-itertools==8.5.0     # via pytest
numpy==1.19.2             # via jplephem, skyfield
packaging==20.4           # via pytest
parso==0.7.1              # via jedi
pexpect==4.8.0            # via ipython
pickleshare==0.7.5        # via ipython
pip-tools==5.3.1          # via -r requirements.in
pluggy==0.13.1            # via pytest
prompt-toolkit==3.0.7     # via ipython
ptyprocess==0.6.0         # via pexpect
py==1.9.0                 # via pytest
pycalverter==1.6.1        # via workalendar
pygments==2.7.1           # via ipython
pyluach==1.1.1            # via workalendar
pyparsing==2.4.7          # via packaging
pytest==6.0.2             # via -r requirements.in
python-dateutil==2.8.1    # via -r requirements.in, freezegun, workalendar
pytz==2020.1              # via workalendar
requests==2.24.0          # via -r requirements.in
sgp4==2.12                # via skyfield
six==1.15.0               # via packaging, pip-tools, python-dateutil
skyfield-data==1.1.0      # via workalendar
skyfield==1.29            # via workalendar
toml==0.10.1              # via pytest
traitlets==5.0.4          # via ipython
urllib3==1.25.10          # via requests
wcwidth==0.2.5            # via prompt-toolkit
workalendar==11.0.1       # via -r requirements.in
zipp==3.2.0               # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
